Update mktest script to copy template files

+++ mktest.sh
@@ -7,6 +7,11 @@ else
 	DIR="."
 fi
 
+if [ ! -d "$DIR/files/test" ]; then
+	echo "Did not finc ./files/test directory."
+	exit 1
+fi
+
 if [ -d "$DIR/test" ]; then
 	rm -r "$DIR/test/"*
 
@@ -16,23 +21,4 @@ if [ -d "$DIR/test" ]; then
 	fi
 fi
 
-mkdir -p "$DIR/test"
-mkdir -p "$DIR/test/docker"
-
-touch "$DIR/test/docker-compose.yml"
-touch "$DIR/test/docker-compose.override.yml"
-touch "$DIR/test/docker-compose.prod.yml"
-touch "$DIR/test/docker-compose.dev.yml"
-
-touch "$DIR/test/.env"
-touch "$DIR/test/.env.local"
-touch "$DIR/test/.env.dev"
-touch "$DIR/test/.env.dev.local"
-touch "$DIR/test/.env.prod"
-touch "$DIR/test/.env.prod.local"
-
-touch "$DIR/test/docker-compose.top.yml"
-touch "$DIR/test/docker/docker-compose.sub.yml"
-
-touch "$DIR/test/.env.top"
-touch "$DIR/test/docker/.env.sub"
+cp -r "$DIR/files/test/" "$DIR/test/"
